requirements.txt
asgiref==3.8.1
Django~=5.2.0
django-cleanup==9.0.0
django-ckeditor==6.7.1
django-environ==0.11.2
mysqlclient==2.2.4
Pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
django-simpleui==2024.7.27
